To whoever made the charts package: How do you get Eclipse to compile javafx packages?
I'm still working on the quota package. Any current implementation is
likely to change for at least the next day or two. I designed a panel
for viewing quota that will display weekly user progress in
JProgressBars. There is an edit button that will show a small window for
editing goals. Currently, nothing in the panel has been implemented and
there is only a basic layout. 

There are a couple known issues with the data package. There is no
support for multiple users yet. Also, the dating scheme may not be ideal
when graphing data from the same month in different years. For example
when graphing data from December 25, 2012 and December 25, 2013 both
dates will be plotted at the X=12 no matter what. When testing with the
JMathPlot library in the resources folder, there was no way to get
around this. However, with JavaFX there may be a way. Regardless, I'd
like to establish standards and requirements for the data and quota
packages with whoever is writing classes that are dependent upon them.