I'm almost done changing the data library to have multiple users. There were a lot of changes so I saved it under org.pht.user.data and the user-related classes under org.pht.user. The new version implements the Serializable interface and will be saved on a single file no matter how many users there are.
The Users class is where loading and saving the file will happen. Just
call the constructor and it will load (or create a file if needed). When
exiting the program call the save() method.

The different users are stored in a HashMap with user names as keys for
their corresponding data. Once a specific user's data is retrieved, the
user has three "sections" of data. These are Quota, DailyLogs, and
Timers (Mao, let me know what you need). 

DailyLogs include health indicators and activities. These are stored in
a HashMap with Calendar objects as keys. Quota is just a wrapper for the
user's goals and I plan to implement the functions to calculate their
progress in the User class.